refactor: change structure of cart.ts
