+++ src/ts/cart-page/cartTypes.ts
@@ -1,0 +1,11 @@
+export interface CartLocalStor {
+    id: number;
+    count: number;
+    price: number;
+  }
+
+export interface AllPromoCodes {
+  id: string;
+  name: string;
+  disc: string;
+}
