+++ src/ts/cart-page/promoCodes.ts
@@ -1,0 +1,6 @@
+import { AllPromoCodes } from "./cartTypes";
+
+export const availablePromoCodes: AllPromoCodes[] = [
+  { id: "RS", name: "Rolling Scopes School", disc: "10" },
+  { id: "EPM", name: "EPAM Systems", disc: "10" },
+];
